% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{embedMd}
\alias{embedMd}
\title{embed a .md file, and change title level}
\usage{
embedMd(subfile, addlevel = 0)
}
\arguments{
\item{subfile}{the file you want to embed}

\item{addlevel}{how many levels you want to add}
}
\value{
text
}
\description{
embed a .md file, and change title level
}
