% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{getMd}
\alias{getMd}
\title{like embedMd, but can fetch part of it}
\usage{
getMd(subfile, startText = "", endText = "", withStart = TRUE, addlevel = 0)
}
\arguments{
\item{subfile}{the file you want to embed}

\item{startText}{startText}

\item{endText}{endText}

\item{withStart}{withStart}

\item{addlevel}{how many levels you want to add}
}
\value{
text
}
\description{
like embedMd, but can fetch part of it
}
