% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getPart}
\alias{getPart}
\title{get part of content}
\usage{
getPart(content, startText = "", endText = "", withStart = TRUE)
}
\arguments{
\item{content}{all contents}

\item{startText}{startText}

\item{endText}{endText}

\item{withStart}{is output with startText?}
}
\value{
text
}
\description{
get part of content
}
