% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{rmParts}
\alias{rmParts}
\title{remove part all in content}
\usage{
rmParts(content, startText = "\\\\*\\\\*示例\\\\*\\\\*", endText = "### ")
}
\arguments{
\item{content}{all contents}

\item{startText}{startText}

\item{endText}{endText}
}
\value{
text
}
\description{
remove part all in content
}
