% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{saveMd}
\alias{saveMd}
\title{Save fetched .md file}
\usage{
saveMd(filename, content)
}
\arguments{
\item{filename}{filename}

\item{content}{filecontent}
}
\value{
text
}
\description{
Save fetched .md file
}
