% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikijs.R
\name{wikijsLogin}
\alias{wikijsLogin}
\title{Login wiki.js}
\usage{
wikijsLogin(userToken, graphqlUrl, schemaUrl)
}
\arguments{
\item{url}{wiki.js graphql url}

\item{token}{user token}
}
\value{
text
}
\description{
Login wiki.js
}
