% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikijs.R
\name{fetchMd}
\alias{fetchMd}
\title{Fetch .md file from wiki.js by graphQL interface}
\usage{
fetchMd(con, Id)
}
\arguments{
\item{con}{connect to graphQL}

\item{Id}{pageId}
}
\value{
text
}
\description{
Fetch .md file from wiki.js by graphQL interface
}
