% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{saveMd}
\alias{saveMd}
\title{Save fetched .md file}
\usage{
saveMd(wikiPath, content)
}
\arguments{
\item{wikiPath}{wiki.js page path}

\item{content}{the page content}
}
\value{
text
}
\description{
Save fetched .md file
}
