% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikijs.R
\name{wikijsLogin}
\alias{wikijsLogin}
\title{Login wiki.js}
\usage{
wikijsLogin(userToken, graphqlUrl, schemaUrl)
}
\arguments{
\item{userToken}{user_token}

\item{graphqlUrl}{wiki.js graphql url}

\item{schemaUrl}{wiki.js graphql schema url}
}
\value{
the connect con
}
\description{
get your user_token from https://your_wiki.js_url/graphql:
copy the following codes to graphql and get the long jwt user_token:
mutation {
  authentication {
    login (
      username: "your username(email)"
      password: "your password"
      strategy: "local"
    ) {jwt}
  }
}
}
